python/psapi-test/requirements.txt
numpy >= 2.1.2
opencv-python >= 4.10.0.84
pynose >= 1.5.3